Fixed bug with subqueries returning null during the analyze stage

+++ test/subquery_scalar_noncorrelated.sql
@@ -1,4 +1,4 @@
-create table planets (name text, moons int4, primary key (name));
+create table planets (name text, moons int8, primary key (name));
 insert into planets (name, moons) values ('Earth', 1), ('Mars', 2);
 
 select name, moons from planets where name = (select 'Mars');
